requirements.txt
asgiref==3.5.2
Django==3.2.15
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
